+++ src/main/java/travel/travelapplication/auth/jwt/JwtService.java
@@ -10,7 +10,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import travel.travelapplication.repository.UserRepository;
+import travel.travelapplication.user.repository.UserRepository;
 
 import java.util.Date;
 import java.util.Optional;
