+++ src/main/java/travel/travelapplication/auth/service/CustomOAuth2UserService.java
@@ -12,8 +12,8 @@ import org.springframework.transaction.annotation.Transactional;
 import travel.travelapplication.auth.CustomOAuth2User;
 import travel.travelapplication.auth.dto.OAuthAttributes;
 import travel.travelapplication.constant.Role;
-import travel.travelapplication.entity.User;
-import travel.travelapplication.repository.UserRepository;
+import travel.travelapplication.user.domain.User;
+import travel.travelapplication.user.repository.UserRepository;
 
 import java.util.Collections;
