--- src/main/java/travel/travelapplication/entity/LikedPlace.java
@@ -1,22 +1,0 @@
-package travel.travelapplication.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.core.mapping.DBRef;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document("LikedPlace")
-@Getter
-public class LikedPlace {
-
-    @Id
-    private ObjectId id;
-
-    @DBRef
-    private Place place;
-
-    public LikedPlace(Place place) {
-        this.place = place;
-    }
-}
