--- src/main/java/travel/travelapplication/entity/Tag.java
@@ -1,20 +1,0 @@
-package travel.travelapplication.entity;
-
-import jakarta.persistence.*;
-import lombok.*;
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.core.mapping.Document;
-
-@Document("Tag")
-@Getter
-public class Tag {
-
-    @Id
-    private ObjectId id;
-
-    private String name;
-
-    public Tag(String name) {
-        this.name = name;
-    }
-}
