+++ src/main/java/travel/travelapplication/place/presentation/TagController.java
@@ -1,0 +1,42 @@
+package travel.travelapplication.place.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.ModelAttribute;
+import org.springframework.web.bind.annotation.PostMapping;
+import travel.travelapplication.auth.CustomOAuth2User;
+import travel.travelapplication.dto.tag.TagDto;
+import travel.travelapplication.place.domain.Tag;
+import travel.travelapplication.user.application.UserService;
+
+import java.util.List;
+
+@Controller
+@RequiredArgsConstructor
+public class TagController {
+
+    private final UserService userService;
+
+    @PostMapping("/tag")
+    public String addUserTag(@ModelAttribute("tag") TagDto tagDto, @AuthenticationPrincipal CustomOAuth2User oAuth2User)
+            throws IllegalAccessException {
+        String email=oAuth2User.getAttribute("email");
+        Tag tag=new Tag(tagDto.getName());
+        userService.addTag(email, tag);
+
+        return null; // return "tag";
+    }
+
+    @GetMapping("/myroom/member/taglist")
+    public String UserTagList(Model model, @AuthenticationPrincipal CustomOAuth2User oAuth2User)
+            throws IllegalAccessException {
+        String email=oAuth2User.getAttribute("email");
+        List<Tag> tagList = userService.findAllTag(email);
+        model.addAttribute("tagList", tagList);
+
+        return null; // return "tagList";
+    }
+}
