+++ src/main/java/travel/travelapplication/plan/application/PlanService.java
@@ -1,0 +1,48 @@
+package travel.travelapplication.plan.application;
+
+import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
+import org.springframework.stereotype.Service;
+import travel.travelapplication.place.domain.Place;
+import travel.travelapplication.plan.domain.Plan;
+import travel.travelapplication.user.domain.UserPlan;
+import travel.travelapplication.plan.repository.PlanRepository;
+
+import java.util.ArrayList;
+import java.util.List;
+
+@Service
+@RequiredArgsConstructor
+@Slf4j
+public class PlanService {
+
+    private final PlanRepository planRepository;
+
+    public List<Plan> searchByPlace(String keyword) {
+        List<Plan> plans = planRepository.findAll();
+        List<Plan> findPlans=new ArrayList<>();
+
+        for(Plan plan : plans) {
+            UserPlan userPlan = plan.getUserPlan();
+            if(findPlaces(userPlan, keyword)) {
+                findPlans.add(plan);
+            }
+        }
+        return findPlans;
+    }
+
+    private boolean findPlaces(UserPlan userPlan, String keyword) {
+        List<Place> places = userPlan.getPlaces();
+
+        for(Place place : places) {
+            if(place.getName().contains(keyword)) {
+                return true;
+            }
+        }
+        return false;
+    }
+
+    public List<Plan> findAll() {
+        return planRepository.findAll();
+    }
+}
