+++ src/main/java/travel/travelapplication/plan/presentation/PlanController.java
@@ -1,0 +1,31 @@
+package travel.travelapplication.plan.presentation;
+
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import travel.travelapplication.plan.domain.Plan;
+import travel.travelapplication.plan.application.PlanService;
+
+import java.util.List;
+
+@Controller
+@RequiredArgsConstructor
+public class PlanController {
+
+    private final PlanService planService;
+
+    @GetMapping("/search")
+    public String planView(Model model, @RequestParam(value = "keyword", required = false) String keyword) {
+
+        if(keyword!=null) {
+            List<Plan> findPlans = planService.searchByPlace(keyword);
+            model.addAttribute("findPlans", findPlans);
+        } else {
+            List<Plan> plans=planService.findAll();
+            model.addAttribute("plans", plans);
+        }
+        return null;
+    }
+}
