--- src/main/java/travel/travelapplication/repository/UserRepository.java
@@ -1,13 +1,0 @@
-package travel.travelapplication.repository;
-
-import org.bson.types.ObjectId;
-import org.springframework.data.mongodb.repository.MongoRepository;
-import travel.travelapplication.entity.User;
-
-import java.util.Optional;
-
-public interface UserRepository extends MongoRepository<User, ObjectId> {
-    Optional<User> findByEmail(String email);
-
-    Optional<User> findByRefreshToken(String refreshToken);
-}
