+++ src/main/java/travel/travelapplication/user/domain/SavedPlan.java
@@ -1,0 +1,24 @@
+package travel.travelapplication.user.domain;
+
+import jakarta.persistence.Id;
+import lombok.*;
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.core.mapping.DBRef;
+import org.springframework.data.mongodb.core.mapping.Document;
+import travel.travelapplication.plan.domain.Plan;
+
+@Document("SavedPlan")
+@Getter
+public class SavedPlan {
+
+    @Id
+    private ObjectId id;
+
+    @DBRef
+    private Plan plan;
+
+    public SavedPlan(Plan plan) {
+        this.plan = plan;
+    }
+
+}
