+++ src/main/java/travel/travelapplication/user/repository/UserPlanRepository.java
@@ -1,0 +1,10 @@
+package travel.travelapplication.user.repository;
+
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import travel.travelapplication.user.domain.UserPlan;
+
+public interface UserPlanRepository extends MongoRepository<UserPlan, ObjectId> {
+
+//    List<UserPlan> findByPlace(String placeName);
+}
