+++ src/main/java/travel/travelapplication/user/repository/UserRepository.java
@@ -1,0 +1,13 @@
+package travel.travelapplication.user.repository;
+
+import org.bson.types.ObjectId;
+import org.springframework.data.mongodb.repository.MongoRepository;
+import travel.travelapplication.user.domain.User;
+
+import java.util.Optional;
+
+public interface UserRepository extends MongoRepository<User, ObjectId> {
+    Optional<User> findByEmail(String email);
+
+    Optional<User> findByRefreshToken(String refreshToken);
+}
